fix(build): Resolve 139 build warnings across codebase (#461)
* fix(build): Resolve 139 build warnings across codebase

## Changes

### Compiler Configuration
- Add `-Xexpect-actual-classes` flag to suppress ~70 expect/actual class warnings (KT-61573)
- Remove obsolete `kotlin.ExperimentalMultiplatformApi` opt-in (no longer exists)
- Make JVM args platform-specific to avoid warnings about non-existent packages

### Deprecated API Fixes
- Update Material Icons to use AutoMirrored variants (InsertDriveFile, OpenInNew, ArrowBack, ViewSidebar)
- Replace deprecated `kotlinx.datetime` APIs with `kotlin.time` equivalents
- Replace deprecated `systemCpuLoad` with `cpuLoad` in PerformanceMonitor
- Update BossWindow to use Compose Multiplatform resource API

### Code Quality Fixes
- Remove unnecessary non-null assertions (!!)
- Remove unnecessary safe calls (?.)
- Remove unnecessary type casts
- Remove redundant else branches in exhaustive when expressions
- Fix condition always true/false warnings
- Add missing @OptIn annotations for experimental APIs
- Suppress deprecation warning for intentionally deprecated function calls

### Platform-Specific JVM Args
- macOS: sun.lwawt, sun.lwawt.macosx packages
- Linux: sun.awt.X11 package
- Eliminates "package not in java.desktop" warnings on non-macOS platforms

## Remaining Warnings (Intentional)
- 2 K1 API warnings in PSIBootstrap.kt (both modules) - left visible as reminder
  for future K2 Analysis API migration when it becomes production-ready


* docs: Comprehensive CLAUDE.md rewrite with fact-checked information

## Updates
- JxBrowser version: 8.8.0 → 8.15.0
- Added bosseditor/ module to architecture
- Removed hardcoded version number (v8.11.4)

## New Sections
- BossEditor Module (LSP, PSI, advanced editor features)
- Performance Monitoring (real-time metrics)
- Dashboard System (start screen cards)
- Download Manager (browser downloads)
- Chromium Branding (custom builds)

## Cleaned Up
- Updated resolved/known issues
- Added additional documentation references
- Moved Git Integration to implemented features


* docs: Restructure CLAUDE.md to improve scannability

- Reduce root CLAUDE.md from 504 to 108 lines (78% smaller)
- Move detailed subsystem docs to docs/SUBSYSTEMS.md
- Move BossEditor module docs to docs/BOSSEDITOR.md
- Move feature docs to docs/FEATURES.md
- Update documentation links in root file


---------

+++ composeApp/src/desktopMain/kotlin/ai/rever/boss/psi/NavigationHandler.kt
@@ -160,7 +160,7 @@ class NavigationHandler(
      * Handle navigation on Cmd+Click.
      */
     private fun handleNavigation(e: MouseEvent) {
-        val offset = textArea.viewToModel2D(e.point).toInt()
+        val offset = textArea.viewToModel2D(e.point)
         if (offset < 0) return
 
         println("[Navigation] Cmd+Click at offset $offset in file: $currentFilePath")
@@ -218,7 +218,7 @@ class NavigationHandler(
      * Handle hover for highlighting.
      */
     private fun handleHover(e: MouseEvent) {
-        val offset = textArea.viewToModel2D(e.point).toInt()
+        val offset = textArea.viewToModel2D(e.point)
         if (offset < 0) {
             highlighter.clearHighlight()
             return
